requirements.txt
alembic==1.0.10
aniso8601==6.0.0
asn1crypto==0.24.0
attrs==19.1.0
autopep8==1.4.4
bcrypt==3.1.6
cffi==1.12.3
Click==7.0
cryptography==2.6.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.4.0
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
Mako==1.0.10
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pyrsistent==0.15.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.3
Werkzeug==0.15.2
